VIAPLAYERINFO

+++ Core/MsgID.h
@@ -876,7 +876,7 @@ namespace SPEP
     const UInt8 OFFQQOPENID         = 0x4B;// 解除QQ群绑定 
     const UInt8 GETQQCLANTALK       = 0x4C;// 接收QQ群聊天记录
     const UInt8 SETVAR              = 0x4D;//设置var
-    const UInt16 VIAPLAYERINFO      = 0x4E;// 查询导入玩家信息
+    const UInt8 VIAPLAYERINFO      = 0x4E;// 查询导入玩家信息
 
 }
